Create convenience wrapper-func for loading data.
Function load_hr_mdata does all the steps required for loading
and mapping out model data to high resolution tile-maps.
It takes lon/lat ranges and a year range as input and returns the
time-averaged high-res model data for the chosen variables. The
cubic-sphere faces are determined automatically and the necessary
ones are loaded and checked for overlap with the chosen region.

The high-res data is optionally reprojected on a coarser resolution
using project_width/project_height parameters, that denote the
number of pixels per model grid-cell. This is highly recommended
for performance.

An additional parameter (add_tile_elevation) optionally loads the
high-res DEM elevation data and regrids it to the lat/lon of the
tile-maps used for the model output, using nearest-neighbor.

The resulting bokeh plots can optionally share properties so that
zooming is shared across subplots.